+++ scripts/src/generate-security-txt.ts
@@ -1,8 +1,10 @@
-import { getRepositoryRootDir, isMain, Log } from './utils.js'
 import { join, resolve } from 'path'
 import { promisify } from 'util'
 import { exec } from 'child_process'
 import { writeFile } from 'fs/promises'
+import { isMain } from './utils/is-main.js'
+import { Log } from './utils/log.js'
+import { getRepositoryRootDir } from './utils/get-repository-root-dir.js'
 
 export const SECURITY_TXT = 'security.txt'
