+++ scripts/src/utils/is-main.ts
@@ -1,0 +1,10 @@
+import { fileURLToPath } from 'url'
+
+/**
+ * isMain(import.meta.url)
+ * https://2ality.com/2022/07/nodejs-esm-main.html
+ */
+export function isMain(importMetaUrl: string) {
+  const modulePath = fileURLToPath(importMetaUrl)
+  return process.argv[1] === modulePath
+}
