+++ scripts/src/utils/object-to-json.ts
@@ -1,0 +1,1 @@
+export const objectToJson = (object: object) => JSON.stringify(object, null, 2)
